HW-7: Done
