+++ animals-config-starter/build.gradle
@@ -1,0 +1,19 @@
+plugins {
+    id 'java'
+}
+
+group = 'org.example'
+version = '0.0.1-SNAPSHOT'
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation group: 'org.springframework.boot', name: 'spring-boot-autoconfigure', version: '2.7.4'
+    implementation group: 'org.springframework.boot', name: 'spring-boot-starter', version: '2.7.4'
+}
+
+tasks.register('prepareKotlinBuildScriptModel') {
+
+}
