+++ animals-config-starter/src/main/java/dto/CharacterEnum.java
@@ -1,0 +1,23 @@
+package dto;
+
+import java.util.List;
+import java.util.Random;
+
+public enum CharacterEnum {
+    SMART,
+    TALKATIVE,
+    TRICKY,
+    WICKED,
+    KIND,
+    SHY,
+    LAZY,
+    SAD;
+
+    private static final List<CharacterEnum> VALUES = List.of(values());
+    private static final int SIZE = VALUES.size();
+    private static final Random RANDOM = new Random();
+
+    public static CharacterEnum randomCharacter() {
+        return VALUES.get(RANDOM.nextInt(SIZE));
+    }
+}
