+++ animals-config-starter/src/main/java/dto/Pet/Pet.java
@@ -1,0 +1,7 @@
+package dto.Pet;
+
+import dto.AbstractAnimal;
+
+public abstract class Pet extends AbstractAnimal{
+    abstract protected String getCare();
+}
