+++ animals-config-starter/src/main/java/dto/Predator/Predator.java
@@ -1,0 +1,7 @@
+package dto.Predator;
+
+import dto.AbstractAnimal;
+
+public abstract class Predator extends AbstractAnimal{
+    abstract protected String hunt();
+}
