+++ animals-config-starter/src/main/java/service/CreateAnimalServicePostProcessor.java
@@ -1,0 +1,20 @@
+package service;
+
+import org.springframework.beans.BeansException;
+import org.springframework.beans.factory.config.BeanPostProcessor;
+import dto.AnimalsEnum;
+
+public class CreateAnimalServicePostProcessor implements BeanPostProcessor {
+    @Override
+    public Object postProcessBeforeInitialization(Object bean, String beanName) throws BeansException {
+        if (bean instanceof CreateAnimalService) {
+            ((CreateAnimalService) bean).setAnimalType(AnimalsEnum.randomAnimal());
+        }
+        return bean;
+    }
+
+    @Override
+    public Object postProcessAfterInitialization(Object bean, String beanName) throws BeansException {
+        return bean;
+    }
+}
