+++ animals-config-starter/src/main/resources/application.properties
@@ -1,0 +1,1 @@
+animalNames=Widget,Squee,Nibbler,Swifty,Ginger,Jiggly,Moonpie,Jaxon,Leeko,Cassie,Piku,Crispin,Sugare,Ponyo,Snortle,Meister,Altai,Paladin,Moura,Marzipan,Waly,Suds,Yung,Cheche,Killy
