+++ app/build.gradle
@@ -8,6 +8,8 @@
 
 plugins {
     // Apply the application plugin to add support for building a CLI application in Java.
+    id 'org.springframework.boot' version '3.2.2'
+    id 'io.spring.dependency-management' version '1.0.11.RELEASE'
     id 'application'
 }
 
@@ -18,24 +20,36 @@ repositories {
 
 dependencies {
     // Use JUnit Jupiter for testing.
-    testImplementation 'org.junit.jupiter:junit-jupiter:5.9.1'
-    testImplementation group: 'org.mockito', name: 'mockito-core', version: '2.1.0'
     implementation group: 'org.springframework', name: 'spring-context', version: '6.1.3'
     implementation group: 'org.springframework', name: 'spring-web', version: '6.1.3'
     implementation group: 'javax.annotation', name: 'javax.annotation-api', version: '1.3.2'
-    testImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: '3.2.2'
+    implementation group: 'org.springframework.boot', name: 'spring-boot-autoconfigure', version: '2.7.4'
+    implementation group: 'org.springframework.boot', name: 'spring-boot-starter', version: '2.7.4'
+    implementation project(':animals-config-starter')
 
 
-    // This dependency is used by the application.
-    implementation 'com.google.guava:guava:31.1-jre'
+    testImplementation 'org.junit.jupiter:junit-jupiter:5.9.1'
+    testImplementation group: 'org.mockito', name: 'mockito-core', version: '2.1.0'
 }
 
 application {
     // Define the main class for the application.
-    mainClass = 'hw.App'
+    mainClass = 'hw3.service.Main'
 }
 
 tasks.named('test') {
     // Use JUnit Platform for unit tests.
     useJUnitPlatform()
+}
+
+jar {
+    manifest {
+        attributes(
+                'Main-Class': "hw3.service.Main"
+        )
+    }
+}
+
+bootJar {
+    enabled = true
 }
