--- app/src/main/java/hw3/dto/Pet/Pet.java
@@ -1,7 +1,0 @@
-package hw3.dto.Pet;
-
-import hw3.dto.AbstractAnimal;
-
-public abstract class Pet extends AbstractAnimal{
-    abstract protected String getCare();
-}
