--- app/src/main/java/hw3/dto/Predator/Predator.java
@@ -1,7 +1,0 @@
-package hw3.dto.Predator;
-
-import hw3.dto.AbstractAnimal;
-
-public abstract class Predator extends AbstractAnimal{
-    abstract protected String hunt();
-}
