--- app/src/main/java/hw3/service/CreateAnimalServicePostProcessor.java
@@ -1,20 +1,0 @@
-package hw3.service;
-
-import org.springframework.beans.BeansException;
-import org.springframework.beans.factory.config.BeanPostProcessor;
-import hw3.dto.AnimalsEnum;
-
-public class CreateAnimalServicePostProcessor implements BeanPostProcessor {
-    @Override
-    public Object postProcessBeforeInitialization(Object bean, String beanName) throws BeansException {
-        if (bean instanceof CreateAnimalService) {
-            ((CreateAnimalService) bean).setAnimalType(AnimalsEnum.randomAnimal());
-        }
-        return bean;
-    }
-
-    @Override
-    public Object postProcessAfterInitialization(Object bean, String beanName) throws BeansException {
-        return bean;
-    }
-}
