+++ app/src/main/java/hw3/service/Main.java
@@ -1,35 +1,19 @@
 package hw3.service;
 
-import hw3.dto.Animal;
-import hw3.repository.AnimalsRepository;
-import hw3.repository.AnimalsRepositoryImpl;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
-import org.springframework.context.annotation.ComponentScan;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.ConfigurableApplicationContext;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
-@ComponentScan(basePackages = "hw3")
+@EnableScheduling
+@SpringBootApplication
 public class Main {
     public static void main(String[] args) {
-        AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext(Main.class);
-        AnimalsRepository animalsRepository = context.getBean(AnimalsRepositoryImpl.class);
+        ConfigurableApplicationContext context = SpringApplication.run(Main.class, args);
 
-        System.out.println("--> Вывод всех бинов, которые находятся в контексте ");
-        for (Object o : context.getBeanDefinitionNames())
-            if (!o.toString().contains("org."))
-                System.out.println(o);
-
-        System.out.println("\n--> Животные, которые родились в високосный год: ");
-        for (String animal : animalsRepository.findLeapYearNames()) {
-            System.out.println(animal);
+        for (String beanDefinitionName : context.getBeanDefinitionNames()) {
+            if (beanDefinitionName.toLowerCase().contains("animal"))
+                System.out.println(beanDefinitionName);
         }
-
-        System.out.println("\n--> Животные, которые старше определенного возраста: ");
-        for (Animal animal : animalsRepository.findOlderAnimal(30)) {
-            System.out.println(animal);
-        }
-
-        System.out.println("\n--> Дубликаты животных: ");
-        animalsRepository.printDuplicate();
-
-        context.close();
     }
 }
